backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.8.4
PyMuPDF
httpx==0.27.0
pandas==2.2.2
python-dotenv==1.0.1
selenium>=4.0.0
webdriver-manager>=4.0.0
tqdm>=4.0.0
